Make MPIR 3.0.0 work with MSVC 2017
In order to compile MPIR 3.0.0 on MSVC 2017 one have to:
1) provide WindowsTargetPlatformVersion higher than Windows 8.1,
2) enable targets in mpir.sln for compilation.

It seems that maintainer of MPIR prefers to build it from msbuild.bat,
so he left mpir.sln unselected inside for users to make selections. In
effect, if you provide targets to be build in command line, then they
won't be built, because they weren't preselected for compilation.

I see no other way to preselect targets in mpir.sln than edyting it
forcefully with python.

Differential Revision: https://phabricator.kde.org/D10127