Fix glib build on MacOS
Summary:
This is needed to build glib since the system libiconv.dylib only
provides iconv_open() and not libiconv_open(). We could build
the GNU libiconv to get libiconv_open(). However, this would
create a libiconv.dylib which will be loaded before the system
one and cause errors in other libraries. Forcing the use of the
native libiconv seems like the better solution here.
Alternatively we could build a static libiconv.a and link that into
glib but that seems unncessary.

Test Plan: glib now builds

Reviewers: vonreth, kfunk

Reviewed By: vonreth

Differential Revision: https://phabricator.kde.org/D15079