Add support for session bus dbus for macOS
This should fix all apps tied to dbus to works, which mean a lot ;)
Please note that the bus session is not killed when the application
starting the service would quit. This mean that you are supposed not
to mix (KDE) applications across multiples sources like brew, macport,
craft or even autonomous KDE dmg.