Change URL to a forked repo with a #420351 bugfix
Bug #420351 was reported to the upstream grantlee maintainer
but has been unresolved for quite a while now. It is causing
kmymoney macOS packages to fail to build and possibly any other
package that uses grantlee, either directly or indirectly via
kdepim.

This commit changes the upstream URL to a fork I created with a
bugfix/workaround.

It will be reverted back to original upstream URL once the
upstream maintainer merges the pull-request that contains said
bugfix or solves the issue otherwise.

BUG: 420351