grantlee: fix macOS packaging issue with a patch
Bug #420351 was reported to the upstream grantlee maintainer
but has been unresolved for quite a while now. It is causing
kmymoney macOS packages to fail to build and possibly any other
package that uses grantlee, either directly or indirectly via
kdepim.

This commit enables a patch that works around the issue on
macOS platform.

CCBUG: 420351