Add libs/gpgmepp blueprint
Since gpgme cannot be built with C++ and Qt bindings for MSVC due
to lack of support in their build system, I wrote a patch that adds
CMake build system for the C++ and Qt bindings. The binding libraries
are built on top of the existing C gpgme library. I tried introducing
CMake builds for gpgme as well, but it turned out to be near impossible.
For this reason this is a standalone blueprint rather than an extension
to the current gpgme blueprint.