Add version.ini for KDE PIM
This is essentially the same as the one of the parent level, with one
small difference: The Git URL. The generic kde URLs do work for most
things (and for repositories predating the Gitlab migration), but they
break the automatic translation fetching. For that to work, the Git URL
actually has to match the canonical path in Gitlab.