Patch QtSvg to be buildable without QtWidgets
Usually that is achieved by not building QtWidgets at all, like we do for
our Android packages currently. However, we do have libs/apps that need
QAction, which is in QtWidgets in Qt 5 (moved to QtGui in Qt 6). With this
patch we can still build QtWidgets but nevertheless avoid it being pulled
into APKs that don't need it.