Build Qt5 without the GTK theme
Craft tries to provide as many libraries as possible in order to be distro independent.
Therefor we also provide glib.
We don't provide gtk.
The qt gtk plugin depends on the system GTK which in turn depends many more system libs including glib.
This results in an unstable application as soon as the GTK theme is loaded.