Add linuxdeploy-plugin-craft
This is a plugin for linuxdeploy to set the correct rpath for all libraries
in the plugins directory. The rpath of plugins has to be relative to the
plugins location unlike the rpath of normal libraries which has to be
relative to the location of the executable dynamically linking the
libraries.