Fix libxml2 Android build
This needs a small patch to work around an incompatibility with the ECM
Android toolchain unfortunately. We need to find a way to get rid of the
Threads::Threads hack there which breaks configure time compile checks
in projects explicitly linking against Threads::Threads. Might become
possible with Qt 6, where this is at least no longer required for Qt
itself.