[QtKeychain] Pin to a commit
This affects only the build against Qt6 as Qt5 builds use stable targets: there was a recent change that broke the build of libQuotioent, so choose a commit before the damage.
Also using master as default target is in general not the best idea as it is not very reliable and hard to debug.