Fix incorrect check for BT scan permissions on android sdkver 30+
This adds a missing section which made QtBluetooth require location
permissions on Android 30 and up, even while this is no longer actually
required for Bluetooth in all cases. There is arguably a missing check
for situations where these permissions are not marked as strictly being
unused for location purposes, but it would seem this is likely to be a
vanishingly small number of cases, and will be trapped by Android and
just fail anyway if used in this manner.