Add blueprint for heaptrack
Most of this is boiler plate, except for the custom wrangling
with the AppRun and appimage entry points. The result is an
appimage that closely resembles the one I build manually so far.
It does seem to contain a lot more data (and is thus larger),
meaning the blacklist might need to be further tweaked in the future.

Sadly, it doesn't seem to be possible to blacklist plugins or
translations...