Remove hardcoded zlib pkgconfig dependency from libpng's .pc file
zlib is part of the NDK and has no .pc file, thus looking up any package
with zlib in its dependency chain will fail to be found via pkgconfig
(despite otherwise working properly).

We have a similar fix in the Freetype blueprint as well already, and this
fixes building mpvqt.