Force a rebuild of the cached QtDeclarative
We only need this for Android/ARM64, where the fix for emoji glyphs
triggered an ABI change in Freetype that affects QtDeclarative (and
for whatever reason only that, and only on ARM64). It's rather severe
there though, as we produce non-starting APKs.

I'm not aware of a more targeted to achieve this than forcing it on
all platforms.