Updating QT6 recipes to match the QT5 ones for KStars and INDI
KStars has been operating in QT5 and so the recipes in QT5-lts branch are up to date, but the ones in the master branch are not.  I worked this week to bring KStars and related packages up to QT6, so this merge request reflects the necessary blueprints changes.  The QT6 branch in kstars is not merged yet, but the goal is to have this ready for it when it does.