Fix libplacebo/ffmpeg builds on NDKv26
This fixes a confusing error related to Vulkan headers when compiling
libplacebo/ffmpeg - for example in mpv - specifically on Android with
NDK v26. The reason for this is some misplaced Vulkan headers in this
NDK, see https://github.com/android/ndk/issues/1974. But in reality we
should just turn off Vulkan decode support entirely for these libraries
as Google recommends not to use these extensions.