breeze: Remove superfluous cmake options
Breeze doesn't look for KF6FrameworkIntegration and sets
WITH_DECORATIONS=OFF unless (UNIX AND NOT APPLE AND NOT ANDROID).
Hence, using those options for Windows or macOS is superfluous.

Note that the "not" in the removed if-condition was wrong. Looks like a
typo when changing the line
   if OsUtils.isWin() or OsUtils.isMac():
to the now removed line with fd760de259caca35c78016e4f70d9a0ba623599c