Also skip libcurl on Android for now
The latest version has a hard dependency on libpsl, which isn't deployable
on Android.

No actual functional change, this was disabled in the CI configuration
in ci-utilities until this morning, but unlike other things in there that
still is needed after all. At least it's consistent between CI and local
builds now.