Adding phonon for MacOS since I got Darwin working as a backend
I had disabled phonon for the MacOS KStars build since VLC was not building properly.  Also on QT5 we had to do a lot to get the VLC backend working and this did not work for QT6 anyway.  This adds phonon back since I got KStars to use "darwin" as the backend for phonon for QT6 instead of VLC.